+++ litebird_sim/mapmaking/h_maps.py
@@ -14,7 +14,7 @@ from litebird_sim.healpix import UNSEEN_PIXEL_VALUE
 from litebird_sim.hwp import HWP
 from litebird_sim.maps_and_harmonics import HealpixMap
 from litebird_sim.observations import Observation
-from litebird_sim.pointings_in_obs import (
+from litebird_sim.observation_utilities import (
     _get_hwp_angle,
     _normalize_observations_and_pointings,
 )